+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/MetricItemExtensions.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Collections.Concurrent;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Google.Protobuf;
 using Google.Protobuf.Collections;
@@ -267,37 +268,12 @@ internal static class MetricItemExtensions
                             }
                         }
 
-                        var exemplars = metricPoint.GetExemplars();
-                        foreach (var examplar in exemplars)
+                        if (metricPoint.TryGetExemplars(out var exemplars))
                         {
-                            if (examplar.Timestamp != default)
+                            foreach (ref readonly var exemplar in exemplars)
                             {
-                                byte[] traceIdBytes = new byte[16];
-                                examplar.TraceId?.CopyTo(traceIdBytes);
-
-                                byte[] spanIdBytes = new byte[8];
-                                examplar.SpanId?.CopyTo(spanIdBytes);
-
-                                var otlpExemplar = new OtlpMetrics.Exemplar
-                                {
-                                    TimeUnixNano = (ulong)examplar.Timestamp.ToUnixTimeNanoseconds(),
-                                    TraceId = UnsafeByteOperations.UnsafeWrap(traceIdBytes),
-                                    SpanId = UnsafeByteOperations.UnsafeWrap(spanIdBytes),
-                                    AsDouble = examplar.DoubleValue,
-                                };
-
-                                if (examplar.FilteredTags != null)
-                                {
-                                    foreach (var tag in examplar.FilteredTags)
-                                    {
-                                        if (OtlpKeyValueTransformer.Instance.TryTransformTag(tag, out var result))
-                                        {
-                                            otlpExemplar.FilteredAttributes.Add(result);
-                                        }
-                                    }
-                                }
-
-                                dataPoint.Exemplars.Add(otlpExemplar);
+                                dataPoint.Exemplars.Add(
+                                    ToOtlpExemplar(exemplar.DoubleValue, in exemplar));
                             }
                         }
 
@@ -379,51 +355,48 @@ internal static class MetricItemExtensions
         }
     }
 
-    /*
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static OtlpMetrics.Exemplar ToOtlpExemplar(this IExemplar exemplar)
+    private static OtlpMetrics.Exemplar ToOtlpExemplar<T>(T value, in Metrics.Exemplar exemplar)
+        where T : struct
     {
-        var otlpExemplar = new OtlpMetrics.Exemplar();
-
-        if (exemplar.Value is double doubleValue)
+        var otlpExemplar = new OtlpMetrics.Exemplar
         {
-            otlpExemplar.AsDouble = doubleValue;
-        }
-        else if (exemplar.Value is long longValue)
-        {
-            otlpExemplar.AsInt = longValue;
-        }
-        else
-        {
-            // TODO: Determine how we want to handle exceptions here.
-            // Do we want to just skip this exemplar and move on?
-            // Should we skip recording the whole metric?
-            throw new ArgumentException();
-        }
-
-        otlpExemplar.TimeUnixNano = (ulong)exemplar.Timestamp.ToUnixTimeNanoseconds();
-
-        // TODO: Do the TagEnumerationState thing.
-        foreach (var tag in exemplar.FilteredTags)
-        {
-            otlpExemplar.FilteredAttributes.Add(tag.ToOtlpAttribute());
-        }
+            TimeUnixNano = (ulong)exemplar.Timestamp.ToUnixTimeNanoseconds(),
+        };
 
         if (exemplar.TraceId != default)
         {
             byte[] traceIdBytes = new byte[16];
             exemplar.TraceId.CopyTo(traceIdBytes);
-            otlpExemplar.TraceId = UnsafeByteOperations.UnsafeWrap(traceIdBytes);
-        }
 
-        if (exemplar.SpanId != default)
-        {
             byte[] spanIdBytes = new byte[8];
             exemplar.SpanId.CopyTo(spanIdBytes);
+
+            otlpExemplar.TraceId = UnsafeByteOperations.UnsafeWrap(traceIdBytes);
             otlpExemplar.SpanId = UnsafeByteOperations.UnsafeWrap(spanIdBytes);
+        }
+
+        if (typeof(T) == typeof(long))
+        {
+            otlpExemplar.AsInt = (long)(object)value;
+        }
+        else if (typeof(T) == typeof(double))
+        {
+            otlpExemplar.AsDouble = (double)(object)value;
+        }
+        else
+        {
+            Debug.Fail("Unexpected type");
+            otlpExemplar.AsDouble = Convert.ToDouble(value);
+        }
+
+        foreach (var tag in exemplar.FilteredTags)
+        {
+            if (OtlpKeyValueTransformer.Instance.TryTransformTag(tag, out var result))
+            {
+                otlpExemplar.FilteredAttributes.Add(result);
+            }
         }
 
         return otlpExemplar;
     }
-    */
 }