+++ test/OpenTelemetry.Tests/Metrics/MetricExemplarTests.cs
@@ -1,23 +1,18 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#nullable enable
+
 using System.Diagnostics;
 using System.Diagnostics.Metrics;
 using OpenTelemetry.Tests;
 using Xunit;
-using Xunit.Abstractions;
 
 namespace OpenTelemetry.Metrics.Tests;
 
 public class MetricExemplarTests : MetricTestsBase
 {
     private const int MaxTimeToAllowForFlush = 10000;
-    private readonly ITestOutputHelper output;
-
-    public MetricExemplarTests(ITestOutputHelper output)
-    {
-        this.output = output;
-    }
 
     [Theory]
     [InlineData(MetricReaderTemporalityPreference.Cumulative)]
@@ -33,15 +28,21 @@ public class MetricExemplarTests : MetricTestsBase
         using var container = this.BuildMeterProvider(out var meterProvider, builder => builder
             .AddMeter(meter.Name)
             .SetExemplarFilter(new AlwaysOnExemplarFilter())
+            .AddView(
+                "testCounter",
+                new MetricStreamConfiguration
+                {
+                    ExemplarReservoirFactory = () => new SimpleFixedSizeExemplarReservoir(3),
+                })
             .AddInMemoryExporter(exportedItems, metricReaderOptions =>
             {
                 metricReaderOptions.TemporalityPreference = temporality;
             }));
 
-        var measurementValues = GenerateRandomValues(10);
+        var measurementValues = GenerateRandomValues(2, false, null);
         foreach (var value in measurementValues)
         {
-            counter.Add(value);
+            counter.Add(value.Value);
         }
 
         meterProvider.ForceFlush(MaxTimeToAllowForFlush);
@@ -49,14 +50,9 @@ public class MetricExemplarTests : MetricTestsBase
         Assert.NotNull(metricPoint);
         Assert.True(metricPoint.Value.StartTime >= testStartTime);
         Assert.True(metricPoint.Value.EndTime != default);
-        var exemplars = GetExemplars(metricPoint.Value);
 
-        // TODO: Modify the test to better test cumulative.
-        // In cumulative, where SimpleFixedSizeExemplarReservoir's size is
-        // more than the count of new measurements, it is possible
-        // that the exemplar value is for a measurement that was recorded in the prior
-        // cycle. The current ValidateExemplars() does not handle this case.
-        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, measurementValues, false);
+        var exemplars = GetExemplars(metricPoint.Value);
+        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, measurementValues);
 
         exportedItems.Clear();
 
@@ -64,12 +60,11 @@ public class MetricExemplarTests : MetricTestsBase
         Thread.Sleep(10); // Compensates for low resolution timing in netfx.
 #endif
 
-        measurementValues = GenerateRandomValues(10);
-        foreach (var value in measurementValues)
+        var secondMeasurementValues = GenerateRandomValues(1, true, measurementValues);
+        foreach (var value in secondMeasurementValues)
         {
-            var act = new Activity("test").Start();
-            counter.Add(value);
-            act.Stop();
+            using var act = new Activity("test").Start();
+            counter.Add(value.Value);
         }
 
         meterProvider.ForceFlush(MaxTimeToAllowForFlush);
@@ -77,12 +72,29 @@ public class MetricExemplarTests : MetricTestsBase
         Assert.NotNull(metricPoint);
         Assert.True(metricPoint.Value.StartTime >= testStartTime);
         Assert.True(metricPoint.Value.EndTime != default);
+
         exemplars = GetExemplars(metricPoint.Value);
-        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, measurementValues, true);
+
+        if (temporality == MetricReaderTemporalityPreference.Cumulative)
+        {
+            // Current design:
+            //  First collect we saw Exemplar A & B
+            //  Second collect we saw Exemplar C but B remained in the reservoir
+            Assert.Equal(2, exemplars.Count);
+            secondMeasurementValues = secondMeasurementValues.Concat(measurementValues.Skip(1).Take(1)).ToArray();
+        }
+        else
+        {
+            Assert.Single(exemplars);
+        }
+
+        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, secondMeasurementValues);
     }
 
-    [Fact]
-    public void TestExemplarsHistogram()
+    [Theory]
+    [InlineData(MetricReaderTemporalityPreference.Cumulative)]
+    [InlineData(MetricReaderTemporalityPreference.Delta)]
+    public void TestExemplarsHistogram(MetricReaderTemporalityPreference temporality)
     {
         DateTime testStartTime = DateTime.UtcNow;
         var exportedItems = new List<Metric>();
@@ -90,18 +102,30 @@ public class MetricExemplarTests : MetricTestsBase
         using var meter = new Meter($"{Utils.GetCurrentMethodName()}");
         var histogram = meter.CreateHistogram<double>("testHistogram");
 
+        var buckets = new double[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+
         using var container = this.BuildMeterProvider(out var meterProvider, builder => builder
             .AddMeter(meter.Name)
             .SetExemplarFilter(new AlwaysOnExemplarFilter())
+            .AddView(
+                "testHistogram",
+                new ExplicitBucketHistogramConfiguration
+                {
+                    Boundaries = buckets,
+                })
             .AddInMemoryExporter(exportedItems, metricReaderOptions =>
             {
-                metricReaderOptions.TemporalityPreference = MetricReaderTemporalityPreference.Delta;
+                metricReaderOptions.TemporalityPreference = temporality;
             }));
 
-        var measurementValues = GenerateRandomValues(10);
+        var measurementValues = buckets
+            /* 2000 is here to test overflow measurement */
+            .Concat(new double[] { 2000 })
+            .Select(b => (Value: b, ExpectTraceId: false))
+            .ToArray();
         foreach (var value in measurementValues)
         {
-            histogram.Record(value);
+            histogram.Record(value.Value);
         }
 
         meterProvider.ForceFlush(MaxTimeToAllowForFlush);
@@ -109,8 +133,9 @@ public class MetricExemplarTests : MetricTestsBase
         Assert.NotNull(metricPoint);
         Assert.True(metricPoint.Value.StartTime >= testStartTime);
         Assert.True(metricPoint.Value.EndTime != default);
+
         var exemplars = GetExemplars(metricPoint.Value);
-        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, measurementValues, false);
+        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, measurementValues);
 
         exportedItems.Clear();
 
@@ -118,11 +143,11 @@ public class MetricExemplarTests : MetricTestsBase
         Thread.Sleep(10); // Compensates for low resolution timing in netfx.
 #endif
 
-        measurementValues = GenerateRandomValues(10);
-        foreach (var value in measurementValues)
+        var secondMeasurementValues = buckets.Take(1).Select(b => (Value: b, ExpectTraceId: true)).ToArray();
+        foreach (var value in secondMeasurementValues)
         {
             using var act = new Activity("test").Start();
-            histogram.Record(value);
+            histogram.Record(value.Value);
         }
 
         meterProvider.ForceFlush(MaxTimeToAllowForFlush);
@@ -130,8 +155,20 @@ public class MetricExemplarTests : MetricTestsBase
         Assert.NotNull(metricPoint);
         Assert.True(metricPoint.Value.StartTime >= testStartTime);
         Assert.True(metricPoint.Value.EndTime != default);
+
         exemplars = GetExemplars(metricPoint.Value);
-        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, measurementValues, true);
+
+        if (temporality == MetricReaderTemporalityPreference.Cumulative)
+        {
+            Assert.Equal(11, exemplars.Count);
+            secondMeasurementValues = secondMeasurementValues.Concat(measurementValues.Skip(1)).ToArray();
+        }
+        else
+        {
+            Assert.Single(exemplars);
+        }
+
+        ValidateExemplars(exemplars, metricPoint.Value.StartTime, metricPoint.Value.EndTime, secondMeasurementValues);
     }
 
     [Fact]
@@ -152,10 +189,14 @@ public class MetricExemplarTests : MetricTestsBase
                 metricReaderOptions.TemporalityPreference = MetricReaderTemporalityPreference.Delta;
             }));
 
-        var measurementValues = GenerateRandomValues(10);
+        var measurementValues = GenerateRandomValues(10, false, null);
         foreach (var value in measurementValues)
         {
-            histogram.Record(value, new("key1", "value1"), new("key2", "value1"), new("key3", "value1"));
+            histogram.Record(
+                value.Value,
+                new("key1", "value1"),
+                new("key2", "value1"),
+                new("key3", "value1"));
         }
 
         meterProvider.ForceFlush(MaxTimeToAllowForFlush);
@@ -167,33 +208,54 @@ public class MetricExemplarTests : MetricTestsBase
         Assert.NotNull(exemplars);
         foreach (var exemplar in exemplars)
         {
-            Assert.NotNull(exemplar.FilteredTags);
-            Assert.Contains(new("key2", "value1"), exemplar.FilteredTags);
-            Assert.Contains(new("key3", "value1"), exemplar.FilteredTags);
+            Assert.NotEqual(0, exemplar.FilteredTags.MaximumCount);
+
+            var filteredTags = exemplar.FilteredTags.ToReadOnlyList();
+
+            Assert.Contains(new("key2", "value1"), filteredTags);
+            Assert.Contains(new("key3", "value1"), filteredTags);
         }
     }
 
-    private static double[] GenerateRandomValues(int count)
+    private static (double Value, bool ExpectTraceId)[] GenerateRandomValues(
+        int count,
+        bool expectTraceId,
+        (double Value, bool ExpectTraceId)[]? previousValues)
     {
         var random = new Random();
-        var values = new double[count];
+        var values = new (double, bool)[count];
         for (int i = 0; i < count; i++)
         {
-            values[i] = random.NextDouble();
+            var nextValue = random.NextDouble();
+            if (values.Any(m => m.Item1 == nextValue)
+                || previousValues?.Any(m => m.Value == nextValue) == true)
+            {
+                i--;
+                continue;
+            }
+
+            values[i] = (nextValue, expectTraceId);
         }
 
         return values;
     }
 
-    private static void ValidateExemplars(Exemplar[] exemplars, DateTimeOffset startTime, DateTimeOffset endTime, double[] measurementValues, bool traceContextExists)
+    private static void ValidateExemplars(
+        IReadOnlyList<Exemplar> exemplars,
+        DateTimeOffset startTime,
+        DateTimeOffset endTime,
+        (double Value, bool ExpectTraceId)[] measurementValues)
     {
-        Assert.NotNull(exemplars);
+        int count = 0;
+
         foreach (var exemplar in exemplars)
         {
             Assert.True(exemplar.Timestamp >= startTime && exemplar.Timestamp <= endTime, $"{startTime} < {exemplar.Timestamp} < {endTime}");
-            Assert.Contains(exemplar.DoubleValue, measurementValues);
-            Assert.Null(exemplar.FilteredTags);
-            if (traceContextExists)
+            Assert.Equal(0, exemplar.FilteredTags.MaximumCount);
+
+            var measurement = measurementValues.FirstOrDefault(v => v.Value == exemplar.DoubleValue);
+            Assert.NotEqual(default, measurement);
+            if (measurement.ExpectTraceId)
             {
                 Assert.NotEqual(default, exemplar.TraceId);
                 Assert.NotEqual(default, exemplar.SpanId);
@@ -203,6 +265,10 @@ public class MetricExemplarTests : MetricTestsBase
                 Assert.Equal(default, exemplar.TraceId);
                 Assert.Equal(default, exemplar.SpanId);
             }
+
+            count++;
         }
+
+        Assert.Equal(measurementValues.Length, count);
     }
 }